use env var interpolation for cyclonedx, refactor report request logic
