+++ spec/fixtures/processor/remote_uri_headers_verbs/expected_report.json
@@ -2,5 +2,5 @@
   "autoCreate": true,
   "projectName": "random_org/random_project",
   "projectVersion": "1",
-  "bom": "eyJib21Gb3JtYXQiOiJDeWNsb25lRFgiLCJzcGVjVmVyc2lvbiI6IjEuMyIsInNlcmlhbE51bWJlciI6InVybjp1dWlkOmE4MzcyZDIxLTA2YmMtNDc3OS1hYjVjLWJlMGEzMDMwMGJjMyIsInZlcnNpb24iOjEsIm1ldGFkYXRhIjp7fSwiY29tcG9uZW50cyI6W119eyJib21Gb3JtYXQiOiJDeWNsb25lRFgiLCJzcGVjVmVyc2lvbiI6IjEuMyIsInNlcmlhbE51bWJlciI6InVybjp1dWlkOmI3NTY0MWJiLWI4ZTgtNGM4Ny05MGIzLTNiMjk0ZDQ3YTg5YiIsInZlcnNpb24iOjEsIm1ldGFkYXRhIjp7fSwiY29tcG9uZW50cyI6W119eyJib21Gb3JtYXQiOiJDeWNsb25lRFgiLCJzcGVjVmVyc2lvbiI6IjEuMyIsInNlcmlhbE51bWJlciI6InVybjp1dWlkOjVmZGU0MzA0LWRmYzAtNDFjNy05YmJlLWEyNWYwNTczM2QxYSIsInZlcnNpb24iOjEsIm1ldGFkYXRhIjp7fSwiY29tcG9uZW50cyI6W119eyJib21Gb3JtYXQiOiJDeWNsb25lRFgiLCJzcGVjVmVyc2lvbiI6IjEuMyIsInNlcmlhbE51bWJlciI6InVybjp1dWlkOmY1ZmU3YzQ5LTAyNjQtNDk1YS04NDczLTMyNDRmMjQyYWI1OCIsInZlcnNpb24iOjEsIm1ldGFkYXRhIjp7fSwiY29tcG9uZW50cyI6W119"
+  "bom": {"bomFormat":"CycloneDX","specVersion":"1.3","serialNumber":"urn:uuid:b75641bb-b8e8-4c87-90b3-3b294d47a89b","version":1,"metadata":{},"components":[]}
 }
