+++ spec/fixtures/processor/remote_uri_headers_verbs/salus.yaml
@@ -5,15 +5,13 @@ reports:
     put:
     headers:
       repo: 'Random Repo'
-    headers_env_var:
-      X-API-Key: 'DUMMY_API_KEY'
+      X-API-Key: '{{DUMMY_API_KEY}}'
   - uri: https://nerv.tk4/salus-report
     format: cyclonedx-json
     post:
     headers:
       repo: 'Random Repo'
-    headers_env_var:
-      X-API-Key: 'DUMMY_API_KEY'
+      X-API-Key: '{{DUMMY_API_KEY}}'
 builds:
   org: "random_org"
   project: "random_project"
